+++ sdk/resourcemanager/appservice/armappservice/response_types.go
@@ -894,6 +894,18 @@ type PlansClientGetServerFarmSKUsResponse struct {
 	Interface any
 }
 
+// PlansClientGetVirtualNetworkIntegrationResponse contains the response from method PlansClient.GetVirtualNetworkIntegration.
+type PlansClientGetVirtualNetworkIntegrationResponse struct {
+	// Swift Virtual Network Contract
+	SwiftVirtualNetwork
+}
+
+// PlansClientGetVirtualNetworkIntegrationsResponse contains the response from method PlansClient.NewGetVirtualNetworkIntegrationsPager.
+type PlansClientGetVirtualNetworkIntegrationsResponse struct {
+	// Collection of Swift Virtual Networks
+	SwiftVirtualNetworkCollection
+}
+
 // PlansClientGetVnetFromServerFarmResponse contains the response from method PlansClient.GetVnetFromServerFarm.
 type PlansClientGetVnetFromServerFarmResponse struct {
 	// Virtual Network information ARM resource.
@@ -1843,14 +1855,14 @@ type WebAppsClientCreateOrUpdateSourceControlSlotResponse struct {
 
 // WebAppsClientCreateOrUpdateSwiftVirtualNetworkConnectionWithCheckResponse contains the response from method WebAppsClient.CreateOrUpdateSwiftVirtualNetworkConnectionWithCheck.
 type WebAppsClientCreateOrUpdateSwiftVirtualNetworkConnectionWithCheckResponse struct {
-	// Swift Virtual Network Contract. This is used to enable the new Swift way of doing virtual network integration.
-	SwiftVirtualNetwork
+	// Swift Virtual Network Contract (Proxy resource)
+	SwiftVirtualNetworkProxy
 }
 
 // WebAppsClientCreateOrUpdateSwiftVirtualNetworkConnectionWithCheckSlotResponse contains the response from method WebAppsClient.CreateOrUpdateSwiftVirtualNetworkConnectionWithCheckSlot.
 type WebAppsClientCreateOrUpdateSwiftVirtualNetworkConnectionWithCheckSlotResponse struct {
-	// Swift Virtual Network Contract. This is used to enable the new Swift way of doing virtual network integration.
-	SwiftVirtualNetwork
+	// Swift Virtual Network Contract (Proxy resource)
+	SwiftVirtualNetworkProxy
 }
 
 // WebAppsClientCreateOrUpdateVnetConnectionGatewayResponse contains the response from method WebAppsClient.CreateOrUpdateVnetConnectionGateway.
@@ -2745,14 +2757,14 @@ type WebAppsClientGetSourceControlSlotResponse struct {
 
 // WebAppsClientGetSwiftVirtualNetworkConnectionResponse contains the response from method WebAppsClient.GetSwiftVirtualNetworkConnection.
 type WebAppsClientGetSwiftVirtualNetworkConnectionResponse struct {
-	// Swift Virtual Network Contract. This is used to enable the new Swift way of doing virtual network integration.
-	SwiftVirtualNetwork
+	// Swift Virtual Network Contract (Proxy resource)
+	SwiftVirtualNetworkProxy
 }
 
 // WebAppsClientGetSwiftVirtualNetworkConnectionSlotResponse contains the response from method WebAppsClient.GetSwiftVirtualNetworkConnectionSlot.
 type WebAppsClientGetSwiftVirtualNetworkConnectionSlotResponse struct {
-	// Swift Virtual Network Contract. This is used to enable the new Swift way of doing virtual network integration.
-	SwiftVirtualNetwork
+	// Swift Virtual Network Contract (Proxy resource)
+	SwiftVirtualNetworkProxy
 }
 
 // WebAppsClientGetTriggeredWebJobHistoryResponse contains the response from method WebAppsClient.GetTriggeredWebJobHistory.
@@ -3949,14 +3961,14 @@ type WebAppsClientUpdateSourceControlSlotResponse struct {
 
 // WebAppsClientUpdateSwiftVirtualNetworkConnectionWithCheckResponse contains the response from method WebAppsClient.UpdateSwiftVirtualNetworkConnectionWithCheck.
 type WebAppsClientUpdateSwiftVirtualNetworkConnectionWithCheckResponse struct {
-	// Swift Virtual Network Contract. This is used to enable the new Swift way of doing virtual network integration.
-	SwiftVirtualNetwork
+	// Swift Virtual Network Contract (Proxy resource)
+	SwiftVirtualNetworkProxy
 }
 
 // WebAppsClientUpdateSwiftVirtualNetworkConnectionWithCheckSlotResponse contains the response from method WebAppsClient.UpdateSwiftVirtualNetworkConnectionWithCheckSlot.
 type WebAppsClientUpdateSwiftVirtualNetworkConnectionWithCheckSlotResponse struct {
-	// Swift Virtual Network Contract. This is used to enable the new Swift way of doing virtual network integration.
-	SwiftVirtualNetwork
+	// Swift Virtual Network Contract (Proxy resource)
+	SwiftVirtualNetworkProxy
 }
 
 // WebAppsClientUpdateVnetConnectionGatewayResponse contains the response from method WebAppsClient.UpdateVnetConnectionGateway.
@@ -4061,6 +4073,12 @@ type WebSiteManagementClientMoveResponse struct {
 	// placeholder for future response values
 }
 
+// WebSiteManagementClientPurgeUnusedVirtualNetworkIntegrationsResponse contains the response from method WebSiteManagementClient.PurgeUnusedVirtualNetworkIntegrations.
+type WebSiteManagementClientPurgeUnusedVirtualNetworkIntegrationsResponse struct {
+	// Response content upon successful deletion of an unused virtual network.
+	Value *PurgedUnusedVirtualNetworkSuccessResponse
+}
+
 // WebSiteManagementClientUpdatePublishingUserResponse contains the response from method WebSiteManagementClient.UpdatePublishingUser.
 type WebSiteManagementClientUpdatePublishingUserResponse struct {
 	// User credentials used for publishing activity.
@@ -4088,6 +4106,12 @@ type WebSiteManagementClientValidateResponse struct {
 type WebSiteManagementClientVerifyHostingEnvironmentVnetResponse struct {
 	// A class that describes the reason for a validation failure.
 	VnetValidationFailureDetails
+}
+
+// WebSiteManagementClientVirtualNetworkIntegrationsResponse contains the response from method WebSiteManagementClient.VirtualNetworkIntegrations.
+type WebSiteManagementClientVirtualNetworkIntegrationsResponse struct {
+	// Swift Virtual Network Contract
+	SwiftVirtualNetwork
 }
 
 // WorkflowRunActionRepetitionsClientGetResponse contains the response from method WorkflowRunActionRepetitionsClient.Get.